refactor: move getMessages
