+++ packages/server-wallet/src/__test-with-peers__/crash-tolerance/crash-tolerance.test.ts
@@ -15,8 +15,9 @@ import {
   peersTeardown,
   peerWallets,
 } from '../../../jest/with-peers-setup-teardown';
+import {getMessages} from '../../message-service/utils';
 import {getChannelResultFor, getPayloadFor, ONE_DAY} from '../../__test__/test-helpers';
-import {expectLatestStateToMatch, getMessages} from '../utils';
+import {expectLatestStateToMatch} from '../utils';
 
 let channelId: string;
 jest.setTimeout(10_000);
